legislink-pro/requirements.txt
google-adk[database]==0.3.0
yfinance==0.2.56
psutil==5.9.5
litellm==1.66.3
google-generativeai==0.8.5
python-dotenv==1.1.0
google-cloud-aiplatform==1.92.0
google-cloud-storage==2.19.0
google-genai==1.14.0
gitpython==3.1.40
google-adk==0.5.0
mysql-connector-python==9.3.0
mysqlclient==2.2.0
Flask-MySQLdb==2.0.0
Flask==3.1.1
Flask-Cors==6.0.1
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1
redis==6.2.0
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
requests==2.31.0
numpy==1.24.3
gunicorn==21.2.0
pytest==7.4.3
pytest-cov==4.1.0
